♻️ Fix all broken tests caused by the migration of the email part checking point abstraction.
